Move poll shifts form partial to a component
Thanks to it, we can move a few helper methods to the component.
