+++ app/views/admin/poll/shifts/new.html.erb
@@ -8,7 +8,7 @@
   </p>
   <%= render "search_officers" %>
 <% else %>
-  <%= render "form" %>
+  <%= render Admin::Poll::Shifts::FormComponent.new(@shift, booth: @booth, officer: @officer) %>
 <% end %>
 
 <div id="shifts">
